ZK-112: Split JournalHeader component into four slots
