+++ web/src/providers/JournalSliceContextProvider.tsx
@@ -1,8 +1,8 @@
-import { SelectAllAction } from '@/components/journal/JournalHeader'
+import { SelectAllAction } from '@/components/journal/ribbon/JournalEntrySelectionActions'
 import { JournalContext } from '@/contexts/JournalContext'
-import { JournalEditorDateViewSymbol, JournalEditorState, JournalSliceContext } from '@/contexts/JournalSliceContext'
+import { JournalEditorState, JournalSliceContext } from '@/contexts/JournalSliceContext'
 import { getJournalEntries } from '@/database/queries'
-import { DateView, JournalEntry, JournalSlice, MonthlyPeriod, WeeklyPeriod } from '@/types/schema'
+import { JournalEntry, JournalSlice } from '@/types/schema'
 import { calculateNetAmount } from '@/utils/journal'
 import { useQuery } from '@tanstack/react-query'
 import { PropsWithChildren, useContext, useEffect, useMemo, useState } from 'react'
